Removed unused includes

+++ Server/Server/StaticHelper.cpp
@@ -6,7 +6,7 @@ bool StaticHelper::CompareStringCaseInsensitive(std::string const& str1, std::st
     if (str1.length() != str2.length())
         return false;
 
-    for (unsigned int i = 0; i < str1.length(); i++)
+    for (size_t i = 0; i < str1.length(); i++)
         if (tolower(str1[i]) != tolower(str2[i]))
             return false;
